refactor

+++ src/main/java/com/mg/app/PharmaSys/model/vente/VenteDetail.java
@@ -16,7 +16,7 @@ public class VenteDetail {
     private Integer id;
 
     @Column(name = "quantite")
-    private Integer quantite;
+    private Double quantite;
 
     @Column(name = "prix_unitaire")
     private Double prixUnitaire;
